requirements.txt
Django==3.2
commitizen==2.17.4
isort==5.8.0
djangorestframework==3.12.4
markdown==3.3.4
django-filter==2.4.0